+++ react/Login.js
@@ -1,109 +1,37 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { graphql } from 'react-apollo'
 import { Button } from 'vtex.styleguide'
 import { injectIntl, intlShape } from 'react-intl'
-import { compose, map, fromPairs, split, tail } from 'ramda' 
+import { compose, map, fromPairs, split, tail } from 'ramda'
 
-import LoginOptions from './components/LoginOptions'
-import EmailVerification from './components/EmailVerification'
-import EmailAndPassword from './components/EmailAndPassword'
-import CodeConfirmation from './components/CodeConfirmation'
-import AccountOptions from './components/AccountOptions'
-import RecoveryPassword from './components/RecoveryPassword'
+import LoginContent from './LoginContent'
 
 import { truncateString } from './utils/truncate-string'
 import ProfileIcon from './images/ProfileIcon'
 import GET_USER_PROFILE from './queries/profile.gql'
 import { translate } from './utils/translate'
-import { steps } from './utils/steps'
-
 import './global.css'
-
-const STEPS = [
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <LoginOptions
-        page="login-options"
-        titleLabel="loginOptions.title"
-        options={['loginOptions.emailVerification', 'loginOptions.emailAndPassword']}
-        onStateChange={func}
-      />
-    )
-  },
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <EmailVerification
-        next={steps.CODE_CONFIRMATION}
-        previous={steps.LOGIN_OPTIONS}
-        isCreatePassword={state.isCreatePassword}
-        email={state.email}
-        onStateChange={func}
-      />
-    )
-  },
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <EmailAndPassword
-        next={steps.ACCOUNT_OPTIONS}
-        previous={steps.LOGIN_OPTIONS}
-        email={state.email}
-        password={state.password}
-        onStateChange={func}
-      />
-    )
-  },
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <CodeConfirmation
-        next={steps.ACCOUNT_OPTIONS}
-        previous={steps.EMAIL_VERIFICATION}
-        email={state.email}
-        code={state.code}
-        onStateChange={func}
-      />
-    )
-  },
-
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <AccountOptions />
-    )
-  },
-  // eslint-disable-next-line
-  (state, func) => {
-    return (
-      <RecoveryPassword
-        next={steps.ACCOUNT_OPTIONS}
-        previous={steps.EMAIL_PASSWORD}
-        email={state.email}
-        onStateChange={func}
-      />
-    )
-  },
-]
 
 /**
  * Set Cookie from the URL name=value and clean up the URL query params.
+ * Used in the OAuth login.
+ *
  * @param {String} url To set cookie and cleaned up.
  */
 const setCookie = (url) => {
   const { accountAuthCookieName, authCookieValue } = compose(
     fromPairs, map(split('=')), split('&'), tail
-  )(url);
+  )(url)
   if (accountAuthCookieName && authCookieValue) {
     const date = new Date()
-    const ONE_DAY = 24*60*60*1000
+    const ONE_DAY = 24 * 60 * 60 * 1000
     date.setTime(date.getTime() + ONE_DAY)
     document.cookie = `${accountAuthCookieName}=${authCookieValue};
       expires=${date.toUTCString()};
       path=/;
       Http-Only=true;`
-    const cleanUrl = url.substring(0, url.indexOf("?"));
+    const cleanUrl = url.substring(0, url.indexOf('?'))
     window.history.replaceState({}, document.title, cleanUrl)
   }
 }
@@ -113,54 +41,84 @@ class Login extends Component {
   static propTypes = {
     /** Intl object*/
     intl: intlShape,
+    data: PropTypes.shape({}).isRequired,
   }
+
+  boxRef_ = React.createRef()
 
   state = {
-    isCreatePassword: false,
-    step: 0,
-    email: '',
-    password: '',
-    code: '',
+    isBoxOpen: false,
   }
 
-  componentDidMount = () => {
+  handleDocumentMouseUp = e => {
+    const { isBoxOpen } = this.state
+
+    const target = e.target
+
+    if (this.boxRef_.current && !this.boxRef_.current.contains(target)) {
+      if (isBoxOpen) {
+        this.setState({
+          isBoxOpen: false,
+        })
+      }
+
+      this.removeListeners()
+    }
+  }
+
+  componentDidMount() {
     if (location.href.indexOf('accountAuthCookieName') > 0) {
       setCookie(location.href)
     }
   }
 
-  handleUpdateState = state => {
-    this.setState(state)
+  componentWillUnmount() {
+    this.removeListeners()
+  }
+
+  removeListeners = () => {
+    document.removeEventListener('mouseup', this.handleDocumentMouseUp)
+  }
+
+  handleProfileIconClick = () => {
+    console.log(this.state)
+    document.addEventListener('mouseup', this.handleDocumentMouseUp)
+
+    this.setState({
+      isBoxOpen: !this.state.isBoxOpen,
+    })
   }
 
   render() {
     const { data: { profile }, intl } = this.props
-
-    const step = profile ? steps.ACCOUNT_OPTIONS : this.state.step
-    const render = STEPS[step](this.state, this.handleUpdateState)
+    const { isBoxOpen } = this.state
 
     return (
       <div className="vtex-login__container flex items-center relative f6 fr">
-        {profile &&
-          (<div>
-            {translate('login.hello', intl)} {profile.firstName || truncateString(profile.email)}
-          </div>)
-        }
+        {profile && (
+          <div className="vtex-login__profile">
+            {translate('login.hello', intl)} {truncateString(profile.firstName) || truncateString(profile.email)}
+          </div>
+        )}
         <Button
           variation="tertiary"
           size="small"
           icon
+          onClick={this.handleProfileIconClick}
         >
           <ProfileIcon />
         </Button>
-        <div className="vtex-login__box absolute right-0 z-max flex flex-colunm">
-          <div className="vtex-login__arrow-up absolute top-0 right-0 shadow-3" />
-          <div className="shadow-3 mt3">
-            <div className="vtex-login__content pa4 flex items-center justify-center relative bg-white">
-              {render}
+        {isBoxOpen && (
+          <div
+            className="vtex-login__box absolute right-0 z-max flex"
+            ref={this.boxRef_}
+          >
+            <div className="vtex-login__arrow-up absolute top-0 right-0 shadow-3" />
+            <div className="shadow-3 mt3">
+              <LoginContent profile={profile} isInitialScreenOptionOnly />
             </div>
           </div>
-        </div>
+        )}
       </div>
     )
   }
@@ -171,6 +129,7 @@ const options = {
     ssr: false,
   }),
 }
+
 export default injectIntl(
   graphql(GET_USER_PROFILE, options)(Login)
 )